pyproject.toml
[tool.poetry]
name = "easy-recipe"
version = "0.1.0"
description = "Projeto Django com receitas"
authors = ["Matheus Pinheiro <mfp186@hotmail.com>"]
readme = "README.md"
packages = [{include = "easy_recipe", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.2.3,<6.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
dj-database-url = ">=3.0.0,<4.0.0"
django-environ = ">=0.12.0,<0.13.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
drf-spectacular = "^0.28.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
